addressed last comment
